pyproject.toml
[tool.poetry]
name = "calculationengine"
version = "0.1.0"
description = ""
authors = ["kasinos <s.kasinos@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^2.0.2"
pint = ">=0.15, <0.18"
djangorestframework = "^3.13.1"
django = "~4.2"
pathlib = "^1.0.1"
numpy = "^1.24.3"
DateTime = "^5.1"
scipy = "^1.10.1"

django-extensions = "^3.2.1"

[tool.poetry.dev-dependencies]
django-extensions = "^3.2.0"
bpython = "^0.22.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
